--- src/main/java/com/gregori/domain/AbstractEntity.java
@@ -1,14 +1,0 @@
-package com.gregori.domain;
-
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-import java.time.ZonedDateTime;
-import java.time.format.DateTimeFormatter;
-
-@Getter
-public class AbstractEntity {
-    private ZonedDateTime createdAt;
-
-    private ZonedDateTime updatedAt;
-}
