--- src/main/java/com/gregori/dto/item/ItemInsertDto.java
@@ -1,16 +1,0 @@
-package com.gregori.dto.item;
-
-import lombok.Getter;
-
-@Getter
-public class ItemInsertDto {
-	private String name;
-	private Long price;
-	private Long inventory;
-
-	public ItemInsertDto(String name, Long price, Long inventory) {
-		this.name = name;
-		this.price = price;
-		this.inventory = inventory;
-	}
-}
