--- src/main/java/com/gregori/dto/item/ItemUpdateDto.java
@@ -1,18 +1,0 @@
-package com.gregori.dto.item;
-
-import lombok.Getter;
-
-@Getter
-public class ItemUpdateDto {
-	private Long id;
-	private String name;
-	private Long price;
-	private Long inventory;
-
-	public ItemUpdateDto(Long id, String name, Long price, Long inventory) {
-		this.id = id;
-		this.name = name;
-		this.price = price;
-		this.inventory = inventory;
-	}
-}
