+++ src/main/java/com/gregori/item/controller/ItemController.java
@@ -1,0 +1,27 @@
+package com.gregori.item.controller;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+import com.gregori.item.dto.ItemResponseDto;
+import com.gregori.item.service.ItemService;
+
+import lombok.RequiredArgsConstructor;
+
+@Controller
+@RequiredArgsConstructor
+@RequestMapping("/items")
+public class ItemController {
+	private final ItemService itemService;
+
+	@GetMapping("/{itemId}")
+	public ResponseEntity<ItemResponseDto> findItemById(@PathVariable Long itemId) {
+		ItemResponseDto itemResponseDto = itemService.findItemById(itemId);
+
+		return ResponseEntity.status(HttpStatus.OK).body(itemResponseDto);
+	}
+}
