+++ src/main/java/com/gregori/item/mapper/ItemMapper.java
@@ -1,0 +1,16 @@
+package com.gregori.item.mapper;
+
+import java.util.List;
+import java.util.Optional;
+
+import org.apache.ibatis.annotations.Mapper;
+
+import com.gregori.item.domain.Item;
+
+@Mapper
+public interface ItemMapper {
+	Long insert(Item item);
+	Long update(Item item);
+	Optional<Item> findById(Long itemId);
+	List<Item> findAllById(List<Long> itemIds);
+}
